DS-1057 - Move Events logic into more reasonable package to avoid unnecessary complication
