+++ site/components/src/main/java/com/visitscotland/brxm/event/EventHstQueryService.java
@@ -1,4 +1,4 @@
-package com.visitscotland.brxm.rest.event;
+package com.visitscotland.brxm.event;
 
 import com.visitscotland.brxm.hippobeans.IndustryEventBSH;
 import com.visitscotland.brxm.hippobeans.TrainingEventBSH;
