+++ site/components/src/main/resources/META-INF/hst-assembly/overrides/spring-plain-rest-api.xml
@@ -21,7 +21,7 @@
         </bean>
         <bean class="org.apache.cxf.jaxrs.lifecycle.SingletonResourceProvider">
           <constructor-arg>
-            <bean class="com.visitscotland.brxm.rest.EventsSearchService"/>
+            <bean class="com.visitscotland.brxm.event.EventsRestResource"/>
           </constructor-arg>
         </bean>
       </list>
